remove gulp, add script
